Stable.

+++ src/global_user/include/global_user/global_user.hpp
@@ -2,7 +2,7 @@
  * @Description: This is a ros_control learning project!
  * @Author: Liu Biao
  * @Date: 2022-09-05 03:24:50
- * @LastEditTime: 2023-03-03 09:34:37
+ * @LastEditTime: 2023-02-26 13:31:27
  * @FilePath: /TUP-Vision-2023-Based/src/global_user/include/global_user/global_user.hpp
  */
 #ifndef GLOBAL_USER_HPP_
@@ -153,7 +153,7 @@ namespace global_user
         double bullet_speed;
         cv::Mat img;
         Eigen::Quaterniond quat;
-        int64_t timestamp; 
+        double timestamp; 
     };
 
     struct GridAndStride
@@ -255,7 +255,7 @@ namespace global_user
     bool setSharedMemory(SharedMemoryParam& shared_memory_param, int id, int image_width = 1280, int image_height = 1024);
     bool getSharedMemory(SharedMemoryParam& shared_memory_param, int id);
     bool destorySharedMemory(SharedMemoryParam& shared_memory_param);
-    bool autoLabel(bool& is_init, cv::Mat &img, ofstream &file, string &path_name, int64_t &timestamp, int &id, int &color, vector<cv::Point2f> &apex2d, cv::Point2i &roi_offset, cv::Size2i &input_size);
+    bool autoLabel(bool& is_init, cv::Mat &img, ofstream &file, string &path_name, double &timestamp, int &id, int &color, vector<cv::Point2f> &apex2d, cv::Point2i &roi_offset, cv::Size2i &input_size);
 } // namespace global_user
 
 #endif
